+++ src/pages/OfferNew/Compare/CompareTable.tsx
@@ -5,7 +5,7 @@ import { Tooltip } from 'new-components/Tooltip'
 import { otherCompanies } from 'pages/OfferNew/Compare/mock'
 import * as React from 'react'
 import { useMediaQuery } from 'react-responsive'
-import { useTextKeys } from 'utils/hooks/useTextKey'
+import { useTextKeys } from 'utils/hooks/useTextKeys'
 import { Checkmark } from '../../../components/icons/Checkmark'
 import { DownArrow } from '../../../components/icons/DownArrow'
 import { HedvigSymbol } from '../../../components/icons/HedvigSymbol'
