+++ src/pages/OfferNew/Introduction/Sidebar/StartDate.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled'
 import { colorsV2 } from '@hedviginsurance/brand'
 import { Tooltip } from 'new-components/Tooltip'
 import * as React from 'react'
-import { useTextKeys } from 'utils/hooks/useTextKey'
+import { useTextKeys } from 'utils/hooks/useTextKeys'
 
 interface Props {
   insuredAtOtherCompany: boolean
