+++ src/pages/OfferNew/Introduction/Sidebar/index.tsx
@@ -3,7 +3,7 @@ import { colorsV2, fonts } from '@hedviginsurance/brand'
 import { InsuranceType, useRemoveDiscountCodeMutation } from 'generated/graphql'
 import { Button, TextButton } from 'new-components/buttons'
 import * as React from 'react'
-import { useTextKeys } from 'utils/hooks/useTextKey'
+import { useTextKeys } from 'utils/hooks/useTextKeys'
 import { formatPostalNumber } from 'utils/postalNumbers'
 import { CompleteOfferData } from '../../types'
 import { getInsuranceType, isMonthlyCostDeduction } from '../../utils'
