+++ src/pages/OfferNew/Introduction/index.tsx
@@ -3,7 +3,7 @@ import styled from '@emotion/styled'
 import { colorsV2 } from '@hedviginsurance/brand'
 import * as React from 'react'
 import { animateScroll } from 'react-scroll'
-import { useTextKeys } from 'utils/hooks/useTextKey'
+import { useTextKeys } from 'utils/hooks/useTextKeys'
 import { DownArrow } from '../../../components/icons/DownArrow'
 import { useDocumentScroll } from '../../../utils/hooks/useDocumentScroll'
 import {
