+++ src/pages/OfferNew/Perils/InsuranceValues/index.tsx
@@ -5,7 +5,7 @@ import { DocumentIcon } from 'components/icons/Document'
 import { InsuranceType } from 'generated/graphql'
 import { Tooltip } from 'new-components/Tooltip'
 import * as React from 'react'
-import { useTextKeys } from 'utils/hooks/useTextKey'
+import { useTextKeys } from 'utils/hooks/useTextKeys'
 import { SubSubHeadingBlack } from '../../components'
 import { getInsurancePDFTextKey, getPrebuyPDFTextKey } from '../../utils'
 import { insuranceValues } from './mock'
